Upgrade project to Vue 3 and Vite
This commit encompasses major changes across the project. It upgrades the project from Vue 2 to Vue 3 and replaces Vue CLI Service with Vite as the build tool. Furthermore, it integrates various plugins to support the new configuration such as Vite Plugin Vuetify and replaces a number of scripts and dependencies in package.json. The structure of several source files is also updated to align with the new Vue 3 standards.